updated svg to graph script
